@@ -1,63 +1,49 @@
 {
-  "name": "hostsbegone",
-  "version": "0.0.1",
-  "description": "Never again use hosts file!",
-  "main": "index.js",
-  "scripts": {
-    "dev": "webpack -d --watch",
-    "build": "webpack -p"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/bisko/hostsbegone.git"
-  },
-  "author": "Biser Perchinkov",
-  "license": "GPLv2",
-  "dependencies": {
-    "async": "^2.0.0-rc.6",
-    "babel-core": "6.9.1",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "6.2.4",
-    "babel-plugin-add-module-exports": "0.2.1",
-    "babel-plugin-lodash": "3.2.0",
-    "babel-plugin-syntax-jsx": "6.8.0",
-    "babel-plugin-transform-class-properties": "6.9.1",
-    "babel-plugin-transform-export-extensions": "6.8.0",
-    "babel-plugin-transform-react-display-name": "6.8.0",
-    "babel-plugin-transform-react-jsx": "6.8.0",
-    "babel-plugin-transform-runtime": "6.9.0",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-stage-2": "6.5.0",
-    "babel-register": "6.9.0",
-    "classnames": "^2.2.5",
-    "eslint": "^3.1.0",
-    "eslint-plugin-react": "^5.2.2",
-    "lodash": "^4.17.2",
-    "native-dns": "github:tutosfaciles48/node-dns",
-    "nconf": "^0.8.4",
-    "node-mac": "^0.1.8",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "react-jsonschema-form": "^0.41.1",
-    "react-redux": "^4.4.5",
-    "react-socket": "^0.1.4",
-    "redux": "^3.5.2",
-    "socket.io": "^1.4.8",
-    "socket.io-client": "^1.4.8",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-plugin-lodash": "^3.2.0",
-    "babel-preset-es2015": "^6.9.0",
-    "css-loader": "^0.26.0",
-    "eslint": "^3.1.0",
-    "eslint-plugin-react": "^5.2.2",
-    "node-sass": "^3.13.0",
-    "react-hot-loader": "^1.3.0",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.3"
-  }
+	"name": "hostsbegone",
+	"version": "0.0.1",
+	"description": "Never again use hosts file!",
+	"main": "start.js",
+	"scripts": {
+		"start": "babel src/start.js",
+		"build": "babel src -d build"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/bisko/hostsbegone.git"
+	},
+	"author": "Biser Perchinkov",
+	"license": "GPLv2",
+	"dependencies": {
+		"async": "^2.0.0-rc.6",
+		"babel-core": "6.26.0",
+		"babel-eslint": "^8.2.2",
+		"babel-loader": "7.1.4",
+		"babel-plugin-add-module-exports": "0.2.1",
+		"babel-plugin-lodash": "3.2.0",
+		"babel-plugin-syntax-jsx": "6.18.0",
+		"babel-plugin-transform-class-properties": "6.24.1",
+		"babel-plugin-transform-export-extensions": "6.22.0",
+		"babel-plugin-transform-react-display-name": "6.25.0",
+		"babel-plugin-transform-react-jsx": "6.24.1",
+		"babel-plugin-transform-runtime": "6.23.0",
+		"babel-preset-es2015": "6.9.0",
+		"babel-preset-stage-2": "6.24.1",
+		"babel-register": "6.26.0",
+		"classnames": "^2.2.5",
+		"eslint": "^4.19.1",
+		"eslint-plugin-react": "^7.7.0",
+		"lodash": "^4.17.2",
+		"native-dns": "github:tutosfaciles48/node-dns",
+		"nconf": "^0.10.0",
+		"node-mac": "^0.1.8",
+		"socket.io": "^2.1.0",
+		"socket.io-client": "^2.1.0"
+	},
+	"devDependencies": {
+		"babel-cli": "^6.26.0",
+		"babel-plugin-lodash": "^3.2.0",
+		"babel-preset-env": "^1.6.1",
+		"babel-preset-es2015": "^6.9.0",
+		"eslint": "^4.19.1"
+	}
 }
